push

+++ module5/assignment/picalc.c
@@ -12,7 +12,7 @@ double f(double a)
 
 int main(int argc,char *argv[])
 {
-    unsigned long int n, myid, numprocs, i, j;
+    long long int n, myid, numprocs, i, j;
     double PI25DT = 3.141592653589793238462643;
     double mypi, pi, h, sum, x;
     double startwtime = 0.0, endwtime;
